requirements-big_query.txt
google-cloud-bigquery==2.3.1
google-cloud-bigquery-storage==2.1.0